fix: resolve all -Wextra sign-compare and missing-field-initializer warnings
Fix sign-compare warnings:
- tt_stl/tests/test_small_vector.cpp: cast kInlineCapacity comparisons
- tt_stl/tt_stl/reflection.hpp: use std::size_t for loop index
- ttnn/cpp/ttnn/operations/functions.hpp: fix iteration variable types
- ttnn/cpp/ttnn/operations/ccl/ccl_common.hpp: cast slice_dim to std::size_t
- tests/tt_metal/test_utils/test_common.hpp: use std::size_t for loop index
- tools/scaleout/validation/utils/cluster_validation_utils.cpp: fix uint32_t vs protobuf int comparisons
- tools/scaleout/factory_system_descriptor/utils.cpp: cast protobuf size() to uint32_t
- tools/scaleout/cabling_generator/cabling_generator.cpp: cast protobuf size() to uint32_t

Fix missing designated field initializers:
- tools/scaleout/validation/utils/cluster_validation_utils.cpp: add .defines = {} to EthernetConfig
- tools/scaleout/cabling_generator/cabling_generator.cpp: add .node_type = "" to Host
- ttnn/cpp/ttnn-nanobind/program_descriptors.cpp: add .unpack_to_dest_mode = {} to ComputeConfigDescriptor
- ttnn/cpp/ttnn-nanobind/pytensor.cpp: add .bottom_up = std::nullopt to DeviceLocalBufferConfig

Suppress -Wextra warnings in third-party TracyClient target via CMake.

+++ tools/scaleout/factory_system_descriptor/utils.cpp
@@ -392,7 +392,7 @@ std::set<PhysicalChannelConnection> validate_fsd_against_gsd_impl(
             // Find host_id for each connection by matching hostname
             uint32_t host_id_a = 0;
             uint32_t host_id_b = 0;
-            for (uint32_t i = 0; i < hosts.size(); ++i) {
+            for (uint32_t i = 0; i < static_cast<uint32_t>(hosts.size()); ++i) {
                 if (hosts[i].hostname() == conn.first.hostname) {
                     host_id_a = i;
                 }
